requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
sentence-transformers
faiss-cpu
pymysql
pymongo
python-jose[cryptography]
passlib
python-dotenv
requests
pydantic-settings
SQLAlchemy
